style: Cleanup
