+++ tests/cases/simple.permissive.yml
@@ -1,0 +1,1 @@
+from: alpine
