+++ tests/cases/springboot-maven.extend.permissive.yml
@@ -1,0 +1,1 @@
+extend: springboot-maven.base.permissive.yml
