+++ tests/infinite_loop/a.yml
@@ -1,1 +1,2 @@
-extend: b.yml
+extend:
+  - b.yml
