+++ tests/infinite_loop/b.yml
@@ -1,1 +1,2 @@
-extend: a.yml
+extend:
+  - a.yml
