+++ tests/infinite_loop/self.yml
@@ -1,1 +1,2 @@
-extend: self.yml
+extend:
+  - self.yml
