+++ tests/infinite_loop/stack_size.yml
@@ -1,1 +1,2 @@
-extend: stack_size_0.yml
+extend:
+  - stack_size_0.yml
