+++ tests/infinite_loop/stack_size_3.yml
@@ -1,1 +1,2 @@
-extend: stack_size_4.yml
+extend:
+  - stack_size_4.yml
