+++ tests/infinite_loop/stack_size_5.yml
@@ -1,1 +1,2 @@
-extend: stack_size_6.yml
+extend:
+  - stack_size_6.yml
