+++ tests/infinite_loop/stack_size_7.yml
@@ -1,1 +1,2 @@
-extend: stack_size_8.yml
+extend:
+  - stack_size_8.yml
