+++ tests/infinite_loop/stack_size_9.yml
@@ -1,1 +1,2 @@
-from: alpine
+from:
+  path: alpine
