+++ tests/lib_integration_test.rs
@@ -2,8 +2,36 @@ use dofigen_lib::*;
 use pretty_assertions_sorted::assert_eq_sorted;
 
 #[test]
+fn yaml_to_dockerfile_empty() {
+    let yaml = "";
+
+    let image: Image = from(yaml.into()).map_err(Error::from).unwrap();
+    let dockerfile: String = generate_dockerfile(&image).unwrap();
+
+    assert_eq_sorted!(
+        dockerfile,
+        r#"# This file is generated by Dofigen v0.0.0
+# See https://github.com/lenra-io/dofigen
+
+# syntax=docker/dockerfile:1.7
+
+# runtime
+FROM scratch AS runtime
+USER 1000:1000
+"#
+    );
+
+    let dockerignore: String = generate_dockerignore(&image);
+
+    assert_eq_sorted!(
+        dockerignore,
+        "# This file is generated by Dofigen v0.0.0\n# See https://github.com/lenra-io/dofigen\n\n"
+    );
+}
+
+#[test]
 #[cfg(feature = "permissive")]
-fn yaml_to_dockerfile() {
+fn yaml_to_dockerfile_complexe() {
     let yaml = r#"
         image: scratch
         builders:
